requirements.txt
Mastodon.py
requests
flask==0.12.2
schedule